backend/requirements.txt
flask
flask-cors
Pillow
torch
torchvision
open-clip-torch
opencv-python
requests
pymongo
dnspython


